Remove old client Regex code, migrate tests to server side
