+++ src/core/client/admin/helpers/index.ts
@@ -1,8 +1,4 @@
 export { default as getQueueConnection } from "./getQueueConnection";
-export {
-  default as getPhrasesRegExp,
-  GetPhrasesRegExpOptions,
-} from "./getPhrasesRegExp";
 export { default as markHTMLNode } from "./markHTMLNode";
 export {
   default as useCommonTranslation,
